requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.2
Pillow==10.1.0